Rename docker_scripts module to docker_squash
