--- docker_scripts/__init__.py
@@ -1,2 +1,0 @@
-from docker_scripts.version import version
-__version__ = version
