+++ docker_squash/__init__.py
@@ -1,0 +1,2 @@
+from docker_squash.version import version
+__version__ = version
