+++ docker_squash/lib/xtarfile.py
