+++ docker_squash/version.py
