+++ tox.ini
@@ -10,5 +10,5 @@ deps=
   pytest-cov
   mock
 commands=
-  py.test -v --cov-report term --cov-report html --cov docker_scripts --basetemp={envtmpdir} --junit-xml target/junit-{envname}.xml --junit-prefix {envname} {posargs}
+  py.test -v --cov-report term --cov-report html --cov docker_squash --basetemp={envtmpdir} --junit-xml target/junit-{envname}.xml --junit-prefix {envname} {posargs}
